% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_gpt_api.R
\name{call_gpt_api}
\alias{call_gpt_api}
\title{Function to send the information and prompt to GPT}
\usage{
call_gpt_api(body)
}
\arguments{
\item{body}{The body of the message, containing both the information and the prompt}
}
\value{
Nothing
}
\description{
This function sends takes care of sending the information and the prompt and is used internally
}
