% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_script.R
\name{current_script}
\alias{current_script}
\title{Collects the current script}
\usage{
current_script()
}
\value{
Nothing
}
\description{
This function gathers the current document and formates it accordingly.
}
