% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_protection.R
\name{data_protection}
\alias{data_protection}
\title{Function for saving the acceptance of the data protection regulations.}
\usage{
data_protection()
}
\value{
Nothing
}
\description{
This function saves whether the person has accepted the data protection regulations as system variable.
}
