% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_path.R
\name{.file_path}
\alias{.file_path}
\title{Function to get the file path}
\usage{
.file_path(setting)
}
\arguments{
\item{setting}{What variable to define, i.e. rainer_dataprot, rainer_logging, rainer_token or rainer_language}
}
\value{
Nothing
}
\description{
This function creates and saves relevant variables necessary for using rainer
}
