% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{.get_config}
\alias{.get_config}
\title{Configuration function}
\usage{
.get_config(setting)
}
\arguments{
\item{setting}{What variable to define, i.e. rainer_dataprot, rainer_logging, rainer_token or rainer_language}
}
\value{
Nothing
}
\description{
This function loads the configuration variables into system variables
}
