% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Function for attaching the package}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Name of library}

\item{pkgname}{Name of package}
}
\value{
Nothing
}
\description{
This function firstly checks if the required information are available and if not, gathers them when attaching the package
}
