% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Function for attaching the package}
\usage{
.onLoad(libname, pkgname)
}
\value{
Nothing
}
\description{
This function firstly checks if the required information are available and if not, gathers them when attaching the package
}
