% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_config.R
\name{.set_config}
\alias{.set_config}
\title{Configuration function}
\usage{
.set_config(setting, content)
}
\arguments{
\item{setting}{What variable to define, i.e. rainer_dataprot, rainer_logging, rainer_token or rainer_language}

\item{content}{How to define the variable, i.e. TRUE/FALSE, german/english or with the token}
}
\value{
Nothing
}
\description{
This function creates and saves relevant variables necessary for using rainer
}
