% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_info.R
\name{environment_info}
\alias{environment_info}
\title{Collect Environment Information}
\usage{
environment_info(error = TRUE)
}
\arguments{
\item{error}{Logical. If TRUE, adds the last error message to the output.}
}
\value{
Nothing
}
\description{
This function gathers information about the current R environment and is used internally.
}
