% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_objects.R
\name{environment_objects}
\alias{environment_objects}
\title{Labels Environment Information}
\usage{
environment_objects()
}
\value{
Nothing
}
\description{
This function labels all loaded objects in the current R environment and is used internally.
}
