% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_language.R
\name{get_language}
\alias{get_language}
\title{Acquires system language}
\usage{
get_language()
}
\value{
Nothing
}
\description{
This function acquires the system language as default for the response language and is used internally
}
