% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_token_validation.R
\name{github_token_validation}
\alias{github_token_validation}
\title{Function for validating the GitHub token}
\usage{
github_token_validation(response_token)
}
\arguments{
\item{response_token}{The token provided by the user which is about to be checked}
}
\value{
Nothing
}
\description{
This function checks whether the provided GitHub token is valid for accessing the LLMs and therefore for using rainer.
}
