% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{header}
\alias{header}
\title{Function to gather the information of data sets}
\usage{
header()
}
\arguments{
\item{body}{The body of the message, containing both the information and the prompt}
}
\value{
Nothing
}
\description{
This function gathers the structure of the loaded datasets and is used internally
}
