% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{language}
\alias{language}
\title{Function for saving the preferred language}
\usage{
language()
}
\value{
Nothing
}
\description{
This function saves the preferred language for the response as a system variable.
}
