% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_initialize.R
\name{log_initialize}
\alias{log_initialize}
\title{Initializing logging settion}
\usage{
log_initialize()
}
\value{
Nothing
}
\description{
This function initializes the logging session if consent to logging is given and saves the session-id
}
