% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_post.R
\name{log_post}
\alias{log_post}
\title{Posting the logged response}
\usage{
log_post(name, content)
}
\arguments{
\item{name}{name of the function which is to be logged}

\item{content}{content to be logged}
}
\value{
Nothing
}
\description{
This function posts the information to be logged
}
