% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logging}
\alias{logging}
\title{Function for saving the acceptance or decline of logging.}
\usage{
logging()
}
\value{
Nothing
}
\description{
This function saves whether the person has accepted or declined logging as a system variable.
}
