% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_activate.R
\name{r_activate}
\alias{r_activate}
\title{Automatic activation of r_error}
\usage{
r_activate(switch = TRUE)
}
\arguments{
\item{switch}{Logical. If TRUE, r_error is triggered automatically when an error message occurs. Use FALSE to deactivate the function.}
}
\value{
Response
}
\description{
This function enables r_error, the explanation of error message, automatically when an error message is
visible.
}
