% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_error.R
\name{r_error}
\alias{r_error}
\title{Provides an explanation for an error message}
\usage{
r_error(error = TRUE, language = NULL)
}
\arguments{
\item{error}{Logical. If TRUE, adds the last error message to the output.}

\item{language}{Language the response is supposed to be in}
}
\value{
Response
}
\description{
This function sends the environment information to GPT and provides feedback on that error
}
