% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_error_addin.R
\name{r_error_addin}
\alias{r_error_addin}
\title{Addin wich r_error, which provides an explanation for an error message}
\usage{
r_error_addin()
}
\value{
Nothing
}
\description{
This function creates the addin in order to execute r_error in the addins menu, used internally
}
