% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_explain.R
\name{r_explain}
\alias{r_explain}
\title{Function to explain something based on a individual prompt}
\usage{
r_explain(
  question =
    "The code does not yield the result I intended. Please explain what the problem might be.",
  error = FALSE,
  language = NULL
)
}
\arguments{
\item{question}{The individual question/prompt which GPT is supposed to answer.}

\item{error}{Logical. If TRUE, adds the last error message to the output.}

\item{language}{Language the response is supposed to be in. Default is the system language.}
}
\value{
Response
}
\description{
This function sends the environment information to GPT and provides feedback on that error based on a prompt
which can be written individually.
}
