% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_explain_addin.R
\name{r_explain_addin}
\alias{r_explain_addin}
\title{Addin for r_explain, allows for individual quetsions in a pop-up}
\usage{
r_explain_addin()
}
\value{
Nothing
}
\description{
This function creates the addin in order to execute r_explain in the addins menu, used internally
}
