% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_export.R
\name{r_export}
\alias{r_export}
\title{Exports the environment information into a JSON file}
\usage{
r_export(filename, error = TRUE)
}
\arguments{
\item{filename}{The name the file is supposed to have}

\item{error}{Logical. If TRUE, adds the last error message to the output.}
}
\value{
nothing
}
\description{
This function exports the gathered information about the environment.
}
