% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_improve.R
\name{r_improve}
\alias{r_improve}
\title{Function to get improvement suggestions}
\usage{
r_improve(error = FALSE)
}
\arguments{
\item{error}{Logical. If TRUE, adds the last error message to the output. FALSE per default.}
}
\value{
Response
}
\description{
Use that function to get improvement suggestions on your current script.
}
