% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_error.R
\name{r_error}
\alias{r_error}
\title{Provides an explanation for an error message}
\usage{
r_error(error = TRUE)
}
\arguments{
\item{error}{Logical. If FALSE, excludes the last error message in the query. TRUE per default.}
}
\value{
Response
}
\description{
Use that function to get an explanation for an error message as well as a suggestion how to correct the issue.
}
