% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_improve.R
\name{r_improve}
\alias{r_improve}
\title{Function to get improvement suggestions}
\usage{
r_improve(error = FALSE, language = NULL)
}
\arguments{
\item{error}{Logical. If TRUE, adds the last error message to the output.}

\item{language}{The language the response is supposed to be in. Default is the system language.}
}
\value{
Response
}
\description{
This function sends the environment information to GPT and provides suggestions on how to improve the script.
}
